Add hook to build *_ui.py files on setup.py install
* Pull pyside2_uic.exe runner out into seperate file for manual use
* Clean up imports to be explicit
* Removed *_ui.py files from source and added *.ui XML files

This change allows only the modifyable code, and not generated code to
be included in source control. It also allows for the original *.ui
files for Qt Creator use to be present for other branches to modify

Calling pip install on the module will compile the *.ui files to *.py
source code, but only in the build target location. This means the
installed module will function as expected, but the compiled *.py files
will not be accessible in the source for development reference or easier
debugging. They may be manually generated by running:
pyside2_uic_runner.py ./ui/ ./src/mailprep/ui/