In progresses interactive file list
Creating an empty file with *.mpjob and then selecting that file
from the Open Job dialog will populate the file list with all files
in that directory and a small properties editor.

The concept was to eventually have this YAML based *.mpjob be
the root of a project folder, and the file list would show files in
that directory that were associated with the project.